+++ src/kraken/analysis/data_manipulation.py
@@ -105,7 +105,9 @@ def examine(
         numeric = (
             False
             if df[column].dtype == "bool"
-            else True if is_numeric_dtype(df[column]) else False
+            else True
+            if is_numeric_dtype(df[column])
+            else False
         )
         if df[column].dtype == "object":
             max_length = df[column].dropna().astype(str).str.len().max()